refactor: share speculative cache rollback
