+++ mlx_vlm/models/deepseek_v4/language.py
@@ -7,6 +7,12 @@ import mlx.nn as nn
 from mlx.nn.layers.distributed import shard_inplace, shard_linear, sum_gradients
 from mlx.utils import tree_flatten
 
+from ...speculative.cache_state import (
+    iter_leaf_caches,
+    needs_replay_snapshot_for_cache,
+    restore_cache_state,
+    snapshot_cache_state,
+)
 from ..base import (
     LanguageModelOutput,
     create_attention_mask,
@@ -1268,179 +1274,6 @@ class DeepseekV4Model(PipelineMixin, nn.Module):
         return self.norm(self.hc_head(h))
 
 
-def _clone_cache_tree(value):
-    if isinstance(value, mx.array):
-        return mx.array(value)
-    if isinstance(value, tuple):
-        return tuple(_clone_cache_tree(v) for v in value)
-    if isinstance(value, list):
-        return [_clone_cache_tree(v) for v in value]
-    if isinstance(value, dict):
-        return {k: _clone_cache_tree(v) for k, v in value.items()}
-    return value
-
-
-def _snapshot_cache_state(
-    caches: List[Any], incoming_tokens: int = 0
-) -> List[Optional[Tuple[Any, Any]]]:
-    return [_snapshot_single_cache(cache, incoming_tokens) for cache in caches]
-
-
-def _needs_replay_snapshot(cache, incoming_tokens: int) -> bool:
-    if cache is None:
-        return False
-    if isinstance(cache, CacheList):
-        return any(
-            _needs_replay_snapshot(child, incoming_tokens) for child in cache.caches
-        )
-    if isinstance(cache, PoolingCache):
-        return int(cache.remainder) + int(incoming_tokens) >= int(cache.ratio)
-    if isinstance(cache, RotatingKVCache):
-        return False
-    return not (hasattr(cache, "trim") and callable(cache.trim))
-
-
-def _needs_replay_snapshot_for_cache(
-    caches: Optional[List[Any]], incoming_tokens: int = 0
-) -> bool:
-    if caches is None:
-        return False
-    return any(_needs_replay_snapshot(cache, incoming_tokens) for cache in caches)
-
-
-def _snapshot_single_cache(cache, incoming_tokens: int = 0):
-    if cache is None:
-        return None
-    if isinstance(cache, CacheList):
-        return (
-            "cache_list",
-            [_snapshot_single_cache(child, incoming_tokens) for child in cache.caches],
-        )
-    if isinstance(cache, PoolingCache):
-        remainder = int(cache.remainder)
-        total = remainder + int(incoming_tokens)
-        overwrite_len = remainder
-        if incoming_tokens > 0:
-            overwrite_len = total % cache.ratio if total >= cache.ratio else 0
-        will_overwrite_remainder = (
-            remainder > 0 and cache.buf_kv is not None and overwrite_len > 0
-        )
-        buf_kv = cache.buf_kv[:, :overwrite_len] if will_overwrite_remainder else None
-        buf_gate = (
-            cache.buf_gate[:, :overwrite_len] if will_overwrite_remainder else None
-        )
-        pooled_len = None if cache.pooled is None else cache.pooled.shape[1]
-        return (
-            "pooling",
-            remainder,
-            _clone_cache_tree(buf_kv),
-            _clone_cache_tree(buf_gate),
-            pooled_len,
-        )
-    if isinstance(cache, RotatingKVCache):
-        return (
-            "rotating",
-            _clone_cache_tree(cache.offset),
-            int(cache._idx),
-            getattr(cache, "start_position", None),
-        )
-    return (
-        "full",
-        _clone_cache_tree(getattr(cache, "state", None)),
-        _clone_cache_tree(getattr(cache, "meta_state", None)),
-    )
-
-
-def _clear_cache_state(cache) -> None:
-    if isinstance(cache, CacheList):
-        for child in cache.caches:
-            _clear_cache_state(child)
-        return
-    if hasattr(cache, "keys"):
-        cache.keys = None
-    if hasattr(cache, "values"):
-        cache.values = None
-    if hasattr(cache, "offset"):
-        cache.offset = 0
-    if hasattr(cache, "_idx"):
-        cache._idx = 0
-    if hasattr(cache, "start_position"):
-        cache.start_position = 0
-    if hasattr(cache, "buf_kv"):
-        cache.buf_kv = None
-    if hasattr(cache, "buf_gate"):
-        cache.buf_gate = None
-    if hasattr(cache, "remainder"):
-        cache.remainder = 0
-    if hasattr(cache, "pooled"):
-        cache.pooled = None
-
-
-def _restore_single_cache(cache, snapshot) -> None:
-    if cache is None or snapshot is None:
-        return
-    kind = snapshot[0]
-    if isinstance(cache, CacheList):
-        for child, child_snapshot in zip(cache.caches, snapshot[1]):
-            _restore_single_cache(child, child_snapshot)
-        return
-    if kind == "pooling":
-        _, remainder, buf_kv, buf_gate, pooled_len = snapshot
-        cache.remainder = int(remainder)
-        if buf_kv is not None:
-            restore_len = int(buf_kv.shape[1])
-            if cache.buf_kv is None or cache.buf_kv.shape[1] < cache.ratio:
-                cache.buf_kv = mx.zeros(
-                    (buf_kv.shape[0], cache.ratio, buf_kv.shape[2]),
-                    dtype=buf_kv.dtype,
-                )
-            if cache.buf_gate is None or cache.buf_gate.shape[1] < cache.ratio:
-                cache.buf_gate = mx.zeros(
-                    (buf_gate.shape[0], cache.ratio, buf_gate.shape[2]),
-                    dtype=buf_gate.dtype,
-                )
-            cache.buf_kv[:, :restore_len] = buf_kv
-            cache.buf_gate[:, :restore_len] = buf_gate
-        if pooled_len is None:
-            cache.pooled = None
-        elif cache.pooled is not None:
-            cache.pooled = cache.pooled[:, :pooled_len]
-        return
-    if kind == "rotating":
-        _, offset, idx, start_position = snapshot
-        cache.offset = _clone_cache_tree(offset)
-        cache._idx = int(idx)
-        if start_position is not None and hasattr(cache, "start_position"):
-            cache.start_position = int(start_position)
-        return
-    _, state, meta_state = snapshot
-    if state is None:
-        _clear_cache_state(cache)
-        return
-    if meta_state is not None and hasattr(type(cache), "meta_state"):
-        cache.meta_state = _clone_cache_tree(meta_state)
-    cache.state = _clone_cache_tree(state)
-
-
-def _restore_cache_state(
-    caches: List[Any], snapshot: List[Optional[Tuple[Any, Any]]]
-) -> None:
-    for cache, entry in zip(caches, snapshot):
-        if cache is None or entry is None:
-            continue
-        _restore_single_cache(cache, entry)
-
-
-def _iter_leaf_caches(caches):
-    for cache in caches:
-        if cache is None:
-            continue
-        if isinstance(cache, CacheList):
-            yield from _iter_leaf_caches(cache.caches)
-        else:
-            yield cache
-
-
 class LanguageModel(nn.Module):
     requires_uniform_batch_acceptance = True
 
@@ -1525,8 +1358,8 @@ class LanguageModel(nn.Module):
     def _speculative_verify(self, inputs: mx.array, cache, sampler=None):
         incoming_tokens = int(inputs.shape[1])
         cache_snapshot = (
-            _snapshot_cache_state(cache, incoming_tokens)
-            if _needs_replay_snapshot_for_cache(cache, incoming_tokens)
+            snapshot_cache_state(cache, incoming_tokens)
+            if needs_replay_snapshot_for_cache(cache, incoming_tokens)
             else None
         )
         sample_logits = sampler is not None
@@ -1574,7 +1407,7 @@ class LanguageModel(nn.Module):
                 raise ValueError(
                     "DeepSeek-V4 speculative rollback requires uniform acceptance."
                 )
-            _restore_cache_state(caches, cache_snapshot)
+            restore_cache_state(caches, cache_snapshot)
             keep = max_a + 1
             if keep > 0:
                 self(verify_inputs[:, :keep], cache=caches, skip_logits=True)
@@ -1585,7 +1418,7 @@ class LanguageModel(nn.Module):
         is_batch = accepted.size > 1
         valid_ends = accepted + 1
 
-        for cache in _iter_leaf_caches(caches):
+        for cache in iter_leaf_caches(caches):
             if trim > 0 and hasattr(cache, "trim"):
                 cache.trim(trim)
             if is_batch and hasattr(cache, "_idx") and max_a > 0:
